+++ src/assets/changelog.md
@@ -2,6 +2,7 @@
 
 - Improve inputs to have labels instead of placeholders.
 - Improve player shuffle logic and design.
+- Add song to selected playlist.
 - Fix issue when adding a new song.
 - Fix player popup is visible when player is expanded.